refactor : 패키지 구조를 좀더 알아보기 쉽게 변경. - domain의 하위로 이동 - blackCustomer와 관련된 model, service, repository를 customer의 하위로 변경
