--- src/main/java/org/prgrms/kdtspringdemo/blacklist/repository/BlackListRepository.java
@@ -1,9 +1,0 @@
-package org.prgrms.kdtspringdemo.blacklist.repository;
-
-import org.prgrms.kdtspringdemo.blacklist.model.BlackCustomer;
-
-import java.util.List;
-
-public interface BlackListRepository {
-    List<BlackCustomer> findAllBlackCustomers();
-}
