+++ src/main/java/org/prgrms/kdtspringdemo/domain/voucher/repository/FileVoucherRepository.java
@@ -1,11 +1,11 @@
-package org.prgrms.kdtspringdemo.voucher.repository;
+package org.prgrms.kdtspringdemo.domain.voucher.repository;
 
+import org.prgrms.kdtspringdemo.domain.voucher.VoucherCreator;
+import org.prgrms.kdtspringdemo.domain.voucher.model.Voucher;
+import org.prgrms.kdtspringdemo.domain.voucher.model.VoucherType;
 import org.prgrms.kdtspringdemo.io.file.CSVReader;
 import org.prgrms.kdtspringdemo.io.file.CSVWriter;
 import org.prgrms.kdtspringdemo.io.file.CsvDto;
-import org.prgrms.kdtspringdemo.voucher.VoucherCreator;
-import org.prgrms.kdtspringdemo.voucher.model.Voucher;
-import org.prgrms.kdtspringdemo.voucher.model.VoucherType;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
