+++ src/main/java/org/prgrms/kdtspringdemo/io/console/Console.java
@@ -1,8 +1,8 @@
 package org.prgrms.kdtspringdemo.io.console;
 
-import org.prgrms.kdtspringdemo.blacklist.model.BlackCustomer;
-import org.prgrms.kdtspringdemo.voucher.model.Voucher;
-import org.prgrms.kdtspringdemo.voucher.model.VoucherType;
+import org.prgrms.kdtspringdemo.domain.customer.model.BlackCustomer;
+import org.prgrms.kdtspringdemo.domain.voucher.model.Voucher;
+import org.prgrms.kdtspringdemo.domain.voucher.model.VoucherType;
 import org.springframework.stereotype.Component;
 
 import java.text.MessageFormat;
